consolidate JSON functionality
- Move JSON-API functions to submodule
- Move `create_json` function to `json` module
